requirements.txt
pandas
numpy
matplotlib
openpyxl
statsmodels
lightgbm
tqdm
jupyter
ipykernel
google-genai
pyarrow
python-dotenv
